[ftio] added the %R conversion to ft_printf functions and changed the # flag on %r to match that new format. the ftio module is now fully compatible to the ansi pedantic norm

+++ libs/ftlib/srcs/ft_freturn.c
@@ -6,13 +6,14 @@
 /*   By: abaurens <abaurens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/04 01:52:23 by abaurens          #+#    #+#             */
-/*   Updated: 2019/03/30 15:51:59 by abaurens         ###   ########.fr       */
+/*   Updated: 2019/09/20 06:32:05 by abaurens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
+#include <inttypes.h>
 
-long int	ft_freturn(void *ptr, long int ret)
+uintptr_t	ft_freturn(void *ptr, uintptr_t ret)
 {
 	free(ptr);
 	return (ret);
